refactor: Restructured tests
