+++ tests/end-to-end/email.e2e.ts
@@ -7,8 +7,9 @@ import {
   LoginPage,
   ProgressDetailsPage,
   ProtectedPage,
-} from './../page-models'
-import { MailClient, getPageUrl, generateRandomEmail } from './../helpers'
+} from './page-models'
+import { MailClient } from './../mocks'
+import { getPageUrl, generateRandomEmail } from './helpers'
 import config from './../config'
 
 fixture`Email campaigns`
