+++ tests/end-to-end/helpers/index.ts
@@ -1,4 +1,3 @@
 export * from './utils'
-export * from './mail'
 export * from './telegram'
 export * from './database'
