+++ tests/end-to-end/page-models/protected.ts
