+++ tests/mocks/index.ts
@@ -1,0 +1,2 @@
+export * from './mail'
+export * from './telegram'
